Revert "Revert "Proposal: Clean up settings.py (#7630)" (#7668)"
This reverts commit 704110fe5278b9c1fc62fb357dfaa685206b3b1b.
